NCI-Agency/anet#379: Further simplify scenario structure
