--- client/tests/sim/scenarios/scenarios/scenario1.js
@@ -1,13 +1,0 @@
-import userTypes from "../users/userTypes"
-import buildup from "../buildup/oldScenario"
-import stories from "../stories/storySim"
-
-// Original scenario which was not in use anymore
-const scenario1 = {
-  description: "Load data into the database",
-  userTypes: userTypes,
-  buildup: buildup,
-  stories: stories
-}
-
-export default scenario1
