--- client/tests/sim/scenarios/scenarios/scenario2.js
@@ -1,12 +1,0 @@
-import userTypes from "../users/userTypes"
-import stories from "../stories/storySim"
-
-// Original scenario which runs stories for ma certain amount of time
-const scenario2 = {
-  description: "Run a set of scenario stories",
-  userTypes: userTypes,
-  buildup: [],
-  stories: stories
-}
-
-export default scenario2
