--- client/tests/sim/scenarios/scenarios/scenario3.js
@@ -1,12 +1,0 @@
-import userTypes from "../users/userTypes"
-import buildup from "../buildup/lotsOfData"
-
-// New scenario intended to load the database with a large amount of data
-const scenario3 = {
-  description: "Load large amounts of data into the database",
-  userTypes: userTypes,
-  buildup: buildup,
-  stories: []
-}
-
-export default scenario3
